Shadow Stack: Finish push_shack. Some bugs in pop_shack.
There's a useless helper function in pop_shack, but without that line,
it will cause segmentation fault. Also, the program never pass the
examination of pop_shack, and thus the pop operation is still not
finished. It may because that it can't see the value of next_eip when
comparing next_eip and sp->guest_eip.